pulito

+++ ft_printf.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mcantell <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: mcantell <mcantell@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/25 17:12:38 by mcantell          #+#    #+#             */
-/*   Updated: 2024/01/29 13:51:09 by mcantell         ###   ########.fr       */
+/*   Updated: 2024/01/30 10:06:52 by mcantell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,13 +24,13 @@ int	ft_proto(va_list args, const char frmt)
 	if (frmt == 's')
 		l += ft_putstr(va_arg(args, char *));
 	if (frmt == 'd' || frmt == 'i')
-		l += ft_printnub(va_arg(args, long int));
+		l += ft_printnub(va_arg(args, long long int));
 	if (frmt == 'u')
 		l += ft_printunsnub(va_arg(args, unsigned int));
 	if (frmt == 'x' || frmt == 'X')
 		l += ft_putnubhex(va_arg(args, unsigned int), frmt);
 	if (frmt == 'p')
-		l += ft_pointer(va_arg(args, size_t));
+		l += ft_pointer(va_arg(args, unsigned long int));
 	return (l);
 }
